Deduplicate no-hoisting test by importing from sibling fixture
Replace copied data.js/reexport.js/index.js in the no-hoisting variant
with a single index.js that imports the sibling self-reexport-star
entry point. Also expand both tests with the failing `import * as Self`
scenario and additional assertions.
