--- turbopack/crates/turbopack-tests/tests/execution/turbopack/exports/self-reexport-star-no-hoisting/input/reexport.js
@@ -1,2 +1,0 @@
-export { Data } from './data'
-export * from './data'
